fix edit

+++ src/main/java/com/fury_cydonian/spring_boot/model/Role.java
@@ -11,7 +11,7 @@ import java.util.Set;
 public class Role implements GrantedAuthority {
     @Id
     @Column(name = "id")
-    private long id;
+    private Long id;
 
     private String name;
 
@@ -22,7 +22,7 @@ public class Role implements GrantedAuthority {
 //            inverseJoinColumns = @JoinColumn(name = "user_id"))
     private Set<User> users;
 
-    public Role(int id, String name) {
+    public Role(Long id, String name) {
         this.id = id;
         this.name = name;
     }
@@ -34,7 +34,7 @@ public class Role implements GrantedAuthority {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
